TECH_flask_chat/flask_chat/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-dotenv==1.0.1
eventlet==0.35.1
# google-generativeai==0.3.2
google-generativeai==0.7.2
# langchain==0.1.17
langchain==0.2.15
#langchain-google-genai==0.0.8
langchain-google-genai==1.0.10
#langchain-community==0.0.36
langchain-community==0.2.7
faiss-cpu==1.11.0